Update Project2_Group17.Rmd
I drew some graphs for response variable and prediction variables but with little comments.
Meanwhile, I fitted GLM model with Poisson family to our data but it's just my opinion about the model. My reason for using Poisson distribution is that time_at_shelter is a continuous variable and for GLM model it should be follow poisson family. 
If you have better choice we could consider about it .